// src/app/precisa/mensajeria/programar-mensaje/model/Dtoprogramaciondetalle.ts
export class Dtoprogramaciondetalle {
    constructor(){

    }
    Id?:                 number;
    IdProgramacion?:     number;
    IdPersona:          any;
    Codigo:             string;
    Nombre:             string;
    Indicador:          number;
    Estado:             number;
    UsuarioCreacion:    string;
    UsuarioModificacion?: string;
    FechaCreacion:       Date;
    FechaModificacion?:   string;
    IpCreacion:          string;
    IpModificacion?:      string;
    DesEstado?:           string;

    num?:                 number;    

}